tests/python/unit/requirements.txt
pandas>=1.0,<1.1
pytest==6.2.1
allure-pytest==2.8.29
Pillow==8.3.2